+++ src/main/java/com/example/restaurant/repository/OrderRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository //Responsble for data access 
-public interface OrderRepo extends JpaRepository<OrderModel, Long> {
+public interface OrderRepo extends JpaRepository<OrderModel, Integer> {
 }
